restore some storybook stories
