+++ src/components/@shared/atoms/Lists/index.stories.tsx
@@ -1,7 +1,15 @@
 import React from 'react'
 import { ComponentStory, ComponentMeta } from '@storybook/react'
 import { ListItem } from '@shared/atoms/Lists'
-import { items } from './index.test'
+
+export const items = [
+  'List item short',
+  'List item long ipsum dolor sit amet, consectetur adipiscing elit. Mauris aliquam facilisis molestie',
+  'List item long ipsum dolor sit amet, consectetur adipiscing elit',
+  'List item short',
+  'List item long ipsum dolor sit amet, consectetur adipiscing elit. Mauris aliquam facilisis molestie',
+  'List item long ipsum dolor sit amet, consectetur adipiscing elit'
+]
 
 export default {
   title: 'Component/@shared/atoms/Lists',
