+++ src/components/@shared/atoms/Table/index.test.tsx
@@ -1,26 +1,12 @@
-import { columns, data } from '../../../../../.jest/__fixtures__/table'
 import { render, screen } from '@testing-library/react'
 import React from 'react'
 import Table, { TableOceanProps } from '.'
-
-export const args: TableOceanProps<any> = { columns, data }
-
-export const argsWithPagination: TableOceanProps<any> = {
-  columns,
-  data: data.flatMap((i) => [i, i, i])
-}
-
-export const argsLoading: TableOceanProps<any> = {
-  isLoading: true,
-  columns: [],
-  data: []
-}
-
-export const argsEmpty: TableOceanProps<any> = {
-  emptyMessage: 'I am empty',
-  columns: [],
-  data: []
-}
+import {
+  args,
+  argsWithPagination,
+  argsLoading,
+  argsEmpty
+} from './index.stories'
 
 describe('Table', () => {
   it('renders without crashing', () => {
